Elaboration becomes a sub-package
