+++ hdl21/elab/context.py
@@ -1,0 +1,9 @@
+"""
+# Elaboration Context 
+"""
+
+
+class Context:
+    """ # Elaboration Context """
+
+    ...  # To be continued!
