+++ hdl21/elab/elaborators/elaborators.py
@@ -1,277 +1,30 @@
-"""
-# Elaboration 
+""" 
+# Elaborator Class(es)
 
-Defines the primary `elaborate` method used to flesh out an in-memory `Module` or `Generator`. 
-Internally defines and uses a number of hierarchical visitor-classes which traverse the hardware hierarchy, 
-performing one or more transformation-passes.  
+Recursive walkers of Hdl21 hierarchies which perform the mechanics of elaboration. 
 """
+
 
 # Std-Lib Imports
 import copy
-from enum import Enum
-from types import SimpleNamespace
 from typing import Union, Any, Dict, List, Optional, Tuple
 
 # PyPi
 from pydantic.dataclasses import dataclass
 
 # Local imports
-from .connect import connectable
-from .module import Module, ExternalModuleCall
-from .instance import InstArray, Instance, PortRef
-from .primitives import PrimitiveCall
-from .bundle import AnonymousBundle, Bundle, BundleInstance, _check_compatible
-from .signal import PortDir, Signal, Visibility, Slice, Concat, Sliceable, NoConn
-from .generator import Generator, GeneratorCall, Default as GenDefault
-from .params import _unique_name
-from .instantiable import Instantiable
+from ...connect import connectable
+from ...module import Module, ExternalModuleCall
+from ...instance import InstArray, Instance, PortRef
+from ...primitives import PrimitiveCall
+from ...bundle import AnonymousBundle, Bundle, BundleInstance, _check_compatible
+from ...signal import PortDir, Signal, Visibility, Slice, Concat, Sliceable, NoConn
+from ...generator import Generator, GeneratorCall, Default as GenDefault
+from ...params import _unique_name
+from ...instantiable import Instantiable
 
-
-class Context:
-    """Elaboration Context"""
-
-    ...  # To be continued!
-
-
-# Type short-hand for elaborate-able types
-Elabable = Union[Module, GeneratorCall]
-# (Plural Version)
-Elabables = Union[Elabable, List[Elabable], SimpleNamespace]
-
-
-def elabable(obj: Any) -> bool:
-    # Function to test this, since `isinstance` doesn't work for `Union`.
-    return isinstance(obj, (Module, Generator, GeneratorCall))
-
-
-class _Elaborator:
-    """ Base Elaborator Class """
-
-    @classmethod
-    def elaborate(cls, top, ctx):
-        """ Elaboration entry-point. Elaborate the top-level object. """
-        return cls(top, ctx).elaborate_top()
-
-    def __init__(self, top: Elabable, ctx: Context):
-        self.top = top
-        self.ctx = ctx
-
-    def elaborate_top(self):
-        """ Elaborate our top node """
-        if not isinstance(self.top, Module):
-            raise TypeError
-        return self.elaborate_module(self.top)
-
-    def elaborate_generator_call(self, call: GeneratorCall) -> Module:
-        """ Elaborate a GeneratorCall """
-        # Only the generator-elaborator can handle generator calls; default it to error on others.
-        raise RuntimeError(f"Invalid call to elaborate GeneratorCall by {self}")
-
-    def elaborate_module(self, module: Module) -> Module:
-        """ Elaborate a Module """
-        # Required for all passes. Defaults to `NotImplementedError`.
-        raise NotImplementedError
-
-    def elaborate_external_module(self, call: ExternalModuleCall) -> ExternalModuleCall:
-        """ Elaborate an ExternalModuleCall """
-        # Default: nothing to see here, carry on
-        return call
-
-    def elaborate_primitive_call(self, call: PrimitiveCall) -> PrimitiveCall:
-        """ Elaborate a PrimitiveCall """
-        # Default: nothing to see here, carry on
-        return call
-
-    def elaborate_bundle_instance(self, inst: BundleInstance) -> None:
-        """ Elaborate an BundleInstance """
-        # Annotate each BundleInstance so that its pre-elaboration `PortRef` magic is disabled.
-        inst._elaborated = True
-
-    def elaborate_bundle(self, bundle: Bundle) -> Bundle:
-        """ Elaborate an Bundle """
-        # Default: nothing to see here, carry on
-        return bundle
-
-    def elaborate_instance_array(self, array: InstArray) -> Instantiable:
-        """ Elaborate an InstArray """
-        # Turn off `PortRef` magic
-        array._elaborated = True
-        # And visit the Instance's target
-        return self.elaborate_instantiable(array._resolved)
-
-    def elaborate_instance(self, inst: Instance) -> Instantiable:
-        """ Elaborate a Module Instance. """
-        # This version of `elaborate_instantiable` is the "post-generators" version used by *most* passes.
-        # The Generator-elaborator is different, and overrides it.
-
-        # Turn off `PortRef` magic
-        inst._elaborated = True
-        # And visit the Instance's target
-        return self.elaborate_instantiable(inst._resolved)
-
-    def elaborate_instantiable(self, of: Instantiable) -> Instantiable:
-        # This version of `elaborate_instantiable` is the "post-generators" version used by *most* passes.
-        # The Generator-elaborator is different, and overrides it.
-        if not of:
-            raise RuntimeError(f"Error elaborating undefined Instance-target {of}")
-        if isinstance(of, Module):
-            return self.elaborate_module(of)
-        if isinstance(of, PrimitiveCall):
-            return self.elaborate_primitive_call(of)
-        if isinstance(of, ExternalModuleCall):
-            return self.elaborate_external_module(of)
-        raise TypeError
-
-    @staticmethod
-    def flatname(
-        segments: List[str], *, avoid: Optional[Dict] = None, maxlen: int = 511
-    ) -> str:
-        """ Create a attribute-name merging string-list `segments`, while avoiding all keys in dictionary `avoid`.
-        Commonly re-used while flattening  nested objects and while creating explicit attributes from implicit ones. 
-        Raises a `RunTimeError` if no such name can be found of length less than `maxlen`. 
-        The default max-length is 511 characters, a value representative of typical limits in target EDA formats. """
-
-        if avoid is None:
-            avoid = {}
-
-        # The default format and result is of the form "seg0_seg1".
-        # If that is in the avoid-keys, append underscores until it's not, or fails.
-        name = "_".join(segments)
-        while True:
-            if len(name) > maxlen:
-                msg = f"Could not generate a flattened name for {segments}: (trying {name})"
-                raise RuntimeError(msg)
-            if name not in avoid:  # Done!
-                break
-            name += "_"  # Collision; append underscore
-        return name
-
-
-class GeneratorElaborator(_Elaborator):
-    """ Hierarchical Generator Elaborator
-    Walks a hierarchy from `top` calling Generators. """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.generator_calls = dict()  # GeneratorCalls to their (Module) results
-        self.modules = dict()  # Module ids to references
-        self.primitive_calls = dict()  # PrimitiveCall ids to references
-        self.ext_module_calls = dict()  # PrimitiveCall ids to references
-
-    def elaborate_top(self):
-        """ Elaborate our top node """
-        if isinstance(self.top, Module):
-            return self.elaborate_module(self.top)
-        if isinstance(self.top, GeneratorCall):
-            return self.elaborate_generator_call(self.top)
-        msg = f"Invalid Elaboration top-level {self.top}, must be a Module or Generator"
-        raise TypeError(msg)
-
-    def elaborate_generator_call(self, call: GeneratorCall) -> Module:
-        """ Elaborate Generator-function-call `call`. Returns the generated Module. """
-
-        # First check out cache
-        if call in self.generator_calls:  # Already done!
-            # Give the `call` a reference to its result.
-            # Note this *has not* necessarily already happened, as the `self.generator_calls` key may be an equally-valued (but distinct) `GeneratorCall`.
-            result = self.generator_calls[call]
-            call.result = result
-            return result
-
-        # Support the "no paramclass constructor" invocation.
-        # Create an instance of the generator's parameter-class if keyword args were provided instead.
-        # If both were provided, fail.
-        if call.kwargs and call.arg is not GenDefault:
-            msg = f"Invalid Generator Call {call}: either provide a single {call.gen.Params} instance or keyword arguments, not both."
-            raise RuntimeError(msg)
-        elif call.arg is GenDefault:
-            # Create an instance of the generator's parameter-class
-            call.arg = call.gen.Params(**call.kwargs)
-
-        # After all that, check that the call has a valid instance of the generator's parameter-class
-        if not isinstance(call.arg, call.gen.Params):
-            msg = f"Invalid Generator Call {call}: {call.gen.Params} instance required, got {call.arg}"
-            raise RuntimeError(msg)
-
-        # The main event: Run the generator-function
-        if call.gen.usecontext:
-            m = call.gen.func(call.arg, self.ctx)
-        else:
-            m = call.gen.func(call.arg)
-
-        # Type-check the result
-        # Generators may return other (potentially nested) generator-calls; unwind any of them
-        while isinstance(m, GeneratorCall):
-            # Note this should hit Python's recursive stack-check if it doesn't terminate
-            m = self.elaborate_generator_call(m)
-        # Ultimately they've gotta resolve to Modules, or they fail.
-        if not isinstance(m, Module):
-            msg = f"Generator {call.gen.func.__name__} returned {type(m)}, must return Module."
-            raise TypeError(msg)
-
-        # Give the GeneratorCall a reference to its result, and store it in our local dict
-        call.result = m
-        self.generator_calls[call] = m
-        # Create a unique name
-        # If the Module that comes back is anonymous, start by giving it a name equal to the Generator's
-        if m.name is None:
-            m.name = call.gen.func.__name__
-        # Then add a unique suffix per its parameter-values
-        m.name += "(" + _unique_name(call.arg) + ")"
-        # Update the Module's `pymodule`, which generally at this point is `hdl21.generator`
-        m._pymodule = call.gen.pymodule
-        # And elaborate the module
-        return self.elaborate_module(m)
-
-    def elaborate_module(self, module: Module) -> Module:
-        """ Elaborate Module `module`. 
-        Primarily performs flattening of Instance Arrays, 
-        and re-connecting to the resultant flattened instances  """
-        if id(module) in self.modules:  # Already done!
-            return module
-
-        if not module.name:
-            msg = f"Anonymous Module {module} cannot be elaborated (did you forget to name it?)"
-            raise RuntimeError(msg)
-
-        # Depth-first traverse instances, ensuring their targets are defined
-        for inst in module.instances.values():
-            self.elaborate_instance(inst)
-        for arr in module.instarrays.values():
-            self.elaborate_instance_array(arr)
-        # Also visit bundle instances, turning off their pre-elab magic
-        for bundle in module.bundles.values():
-            self.elaborate_bundle_instance(bundle)
-
-        # Store a reference to the now-expanded Module in our cache, and return it
-        self.modules[id(module)] = module
-        return module
-
-    def elaborate_instance(self, inst: Instance) -> Instantiable:
-        """ Elaborate a Module Instance. """
-        # This version differs from `Elaborator` in operating on the *unresolved* attribute `inst.of`,
-        # instead of the resolved version `inst._resolved`.
-
-        # Turn off `PortRef` magic
-        inst._elaborated = True
-        # And visit the Instance's target
-        return self.elaborate_instantiable(inst.of)
-
-    def elaborate_instance_array(self, arr: InstArray) -> Instantiable:
-        """ Elaborate an Instance Array. """
-        # This version differs from `Elaborator` in operating on the *unresolved* attribute `inst.of`,
-        # instead of the resolved version `inst._resolved`.
-
-        # Turn off `PortRef` magic
-        arr._elaborated = True
-        # And visit the Instance's target
-        return self.elaborate_instantiable(arr.of)
-
-    def elaborate_instantiable(self, of: Instantiable) -> Instantiable:
-        """ Elaborate an Instance target. Adds the capacity to call `GeneratorCall`s to the more-common base-case. """
-        if isinstance(of, GeneratorCall):
-            return self.elaborate_generator_call(call=of)
-        return super().elaborate_instantiable(of)
+# Import the base class
+from .base import _Elaborator
 
 
 class ImplicitSignals(_Elaborator):
@@ -905,91 +658,6 @@ class BundleConnTypes(_Elaborator):
         return module
 
 
-class ArrayFlattener(_Elaborator):
-    """ Elaboration Pass to Flatten `InstArray`s into `Instance`s, 
-    broadcast and remake their connections. """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.modules = dict()  # Module ids to references
-
-    def elaborate_module(self, module: Module) -> Module:
-        """ Elaborate Module `module`. 
-        Primarily performs flattening of Instance Arrays, 
-        and re-connecting to the resultant flattened instances  """
-        if id(module) in self.modules:  # Already done!
-            return module
-
-        if not module.name:
-            msg = f"Anonymous Module {module} cannot be elaborated (did you forget to name it?)"
-            raise RuntimeError(msg)
-
-        # Flatten Instance arrays
-        while module.instarrays:
-            name, array = module.instarrays.popitem()
-            module.namespace.pop(name)
-            # Visit the array's target
-            target = self.elaborate_instance_array(array)
-
-            # And do the real work: flattening it.
-            if array.n < 1:
-                raise RuntimeError(f"Invalid InstArray {array} with size {array.n}")
-            # Create the new, flat Instances
-            new_insts = []
-            for k in range(array.n):
-                name = self.flatname(
-                    segments=[array.name, str(k)], avoid=module.namespace
-                )
-                inst = module.add(Instance(of=target, name=name))
-                new_insts.append(inst)
-            # And connect them
-            for portname, conn in array.conns.items():
-                if isinstance(conn, BundleInstance):
-                    # All new instances get the same BundleInstance
-                    for inst in new_insts:
-                        inst.connect(portname, conn)
-                elif isinstance(conn, (Signal, Slice, Concat)):
-                    # Get the target-module port, particularly for its width
-                    port = target.ports.get(portname, None)
-                    if not isinstance(port, Signal):
-                        msg = f"Invalid port connection of `{portname}` {port} to {conn} in InstArray {array}"
-                        raise RuntimeError(msg)
-
-                    if port.width == conn.width:
-                        # All new instances get the same signal
-                        for inst in new_insts:
-                            inst.connect(portname, conn)
-                    elif port.width * array.n == conn.width:
-                        # Each new instance gets a slice, equal to its own width
-                        for k, inst in enumerate(new_insts):
-                            slize = conn[k * port.width : (k + 1) * port.width]
-                            if slize.width != port.width:
-                                msg = f"Width mismatch connecting {slize} to {port}"
-                                raise RuntimeError(msg)
-                            inst.connect(portname, slize)
-                    else:  # All other width-values are invalid
-                        msg = f"Invalid connection of {conn} of width {conn.width} to port {portname} on Array {array.name} of width {port.width}. "
-                        msg += f"Valid widths are either {port.width} (broadcasting across instances) and {port.width * array.n} (individually wiring to each)."
-                        raise RuntimeError(msg)
-                elif isinstance(conn, PortRef):
-                    msg = f"Error elaborating {array} in {module}. "
-                    msg += f"Connection {conn} has not been resolved to a `Signal`. "
-                    raise RuntimeError
-                else:
-                    msg = f"Invalid connection to {conn} in InstArray {array}"
-                    raise TypeError(msg)
-
-        # Depth-first traverse instances, ensuring their targets are defined
-        for inst in module.instances.values():
-            self.elaborate_instance(inst)
-        # Also visit bundle instances, turning off their pre-elab magic
-        for bundle in module.bundles.values():
-            self.elaborate_bundle_instance(bundle)
-
-        # Store a reference to the now-expanded Module in our cache, and return it
-        self.modules[id(module)] = module
-        return module
-
 
 class SignalConnTypes(_Elaborator):
     """ Check for connection-type-validity between each Instance and its connections. 
@@ -1051,213 +719,6 @@ class SignalConnTypes(_Elaborator):
             raise RuntimeError(f"Unconnected IO {io} on {inst.name} in {module.name}")
 
 
-class Orphanage(_Elaborator):
-    """ # Orphan-Checking Elaborator Pass 
-
-    Ensures each Module-attribute is "parented" by the `Module` which holds it. 
-    Errant cases can come up for code such as: 
-    
-    ```
-    m1 = h.Module(name='m1')
-    m1.s = h.Signal() # Signal `s` is now "parented" by `m1`
-
-    m2 = h.Module(name='m2')
-    m2.y = m1.s # Now `s` has been "orphaned" (or perhaps "cradle-robbed") by `m2`
-    ```
-
-    This essentially boils down to the difference between Python's native reference-semantics, 
-    and `hdl21.Module`'s notion of "owning" its attributes. 
-    Note other *references* to Module attributes are allowed, such as: 
-
-    ```
-    m1 = h.Module(name='m1')
-    m1.s = h.Signal() # Signal `s` is now "parented" by `m1`
-
-    my_favorite_signals = { "from_m1" : m1.s }
-    ```
-
-    Here the dictionary `my_favorite_signals` retains a reference to Signal `s`. 
-    This does not generate an orphan-error complaint, so long as Module-level ownership is unique and unambiguous. 
-
-    The orphan-test is very simple: each Module-attribute is annotated with a `_parent_module` member 
-    upon insertion into the Module namespace. 
-    Orphan-testing simply requires that for each attribute, this member is identical to the parent Module. 
-    A `RuntimeError` is raised if orphaned attributes are detected. 
-    Otherwise each Module is returned unchanged.
-    """
-
-    def elaborate_module(self, module: Module) -> Module:
-        """ Elaborate a Module """
-        # Check each attribute in the module namespace for orphanage.
-        for attr in module.namespace.values():
-            if attr._parent_module is not module:
-                msg = f"Orphanage: Module {module} attribute {attr} is actually owned by another Module {attr._parent_module}!"
-                raise RuntimeError(msg)
-        # Checks out! Return the module unchanged.
-        return module
-
-
-class SliceResolver(_Elaborator):
-    """ Elaboration pass to resolve slices and concatenations to concrete signals. 
-    Modifies connections to any nested slices, nested concatenations, or combinations thereof. 
-    "Full-width" `Slice`s e.g. `sig[:]` are replaced with their parent `Signal`s.
-
-    TODO: `Slice`s with non-unit `step` are converted to `Concat`s. """
-
-    def elaborate_module(self, module: Module) -> Module:
-        # Depth-first traverse instances, covering their targets first.
-        for inst in module.instances.values():
-            self.elaborate_instance(inst)
-
-        # Then do the real work, updating any necessary connections on each instance
-        for inst in module.instances.values():
-            self.update_instance(module, inst)
-
-        return module
-
-    def update_instance(self, module: Module, inst: Instance) -> None:
-        """ Update all `Slice` and `Concat` valued connections to remove nested `Slice`s """
-
-        for portname, conn in inst.conns.items():
-            if isinstance(conn, (Slice, Concat)):
-                inst.conns[portname] = _resolve_sliceable(conn)
-            # All other connection-types (Signals, Interfaces) are fine
-
-
-def _resolve_sliceable(conn: Sliceable) -> Sliceable:
-    """ Resolve a `Sliceable` to flat-concatenation-amenable elements. """
-    if isinstance(conn, Signal):
-        return conn  # Nothing to do
-    if isinstance(conn, Slice):
-        return _resolve_slice(conn)
-    if isinstance(conn, Concat):
-        return _resolve_concat(conn)
-    raise TypeError(f"Invalid attempt to resolve slicing on {conn}")
-
-
-def _list_slice(slize: Slice) -> List[Slice]:
-    """ Internal recursive helper for `resolve_slice`. 
-    Returns a list of Slices in which each element has a concrete Signal for its parent. """
-
-    # Resolve "full-width" slices to their parent Signals
-    if slize.width == slize.signal.width:
-        # Return a single-element list, after resolution
-        return [_resolve_sliceable(slize.signal)]
-
-    if isinstance(slize.signal, Signal):
-        return [slize]  # Already all good! Just make a one-element list.
-
-    # Do some actual work. Recursively peel off a bit at a time.
-    if slize.width == 1:
-        # Base case: slice is one-bit wide. Reach into the parent signal and grab that bit.
-
-        if isinstance(slize.signal, Slice):
-            parent = slize.signal  # Note this is also a Slice
-            return _list_slice(parent.signal[parent.bot + slize.bot])
-
-        if isinstance(slize.signal, Concat):
-            idx = 0  # Find the `part` including our index
-            for part in slize.signal.parts:
-                if part.width + idx > slize.bot:
-                    return _list_slice(part[slize.bot - idx])
-                idx += part.width
-            msg = f"Slice {slize} is out of bounds of Concat {slize.signal}"
-            raise RuntimeError(msg)
-
-        raise TypeError(f"Invalid attempt to resolve slicing on {slize}")
-
-    # Otherwise recurse in something like a "cons" pattern, splitting between the first bit and the rest.
-
-    if slize.step is not None and slize.step < 0:  # Negative step, begin from `top`
-        first = _list_slice(slize.signal[slize.top])
-        rest = slize.signal[slize.top + slize.step : slize.bot : slize.step]
-        rest = _list_slice(rest)
-
-    else:  # Positive step, begin from `bot`
-        step = slize.step if slize.step is not None else 1
-        first = _list_slice(slize.signal[slize.bot])
-        rest = _list_slice(slize.signal[slize.bot + step : slize.top : slize.step])
-
-    return first + rest
-
-
-def _resolve_slice(slize: Slice) -> Sliceable:
-    """ Resolve a `Slice` to one or more with "concrete" `Signal`s as parents. 
-    
-    Slices of other Slices and Slices of Concats are both valid design-time constructions. 
-    For example: 
-    ```python
-    h.Concat(sig1, sig2, sig3)[1] # Slice of a Concat
-    sig4[0:2][1] # Slice of a Slice
-    ```
-
-    While these may not frequently be created by designers, they are (at least) often created by array broadcasting. 
-    As some point their parents must be resolved to their original Signals, at minimum before export-level name resolution. 
-
-    Resolving Concatenations can generally resolve to more than one Slice, as in: 
-    ```python
-    h.Concat(sig1[0], sig2[0], sig3[0])[0:1] # Requires slices of `sig1` and `sig2`
-    ``` 
-    Such cases create and return a Concatenation. """
-
-    # Break out the slice elements in a list
-    ls = _list_slice(slize)
-    # And convert to either a single element or Concat
-    if len(ls) == 1:  # Resolved to single Slice
-        return ls[0]
-    elif len(ls) > 1:  # Multiple parts required - concatenate them
-        return Concat(*ls)
-
-    raise RuntimeError(f"Error resolving Slice {slize}")
-
-
-def _resolve_concat(conc: Concat) -> Concat:
-    """ Resolve a Concatenation into (a) concrete Signals and (b) Slices of concrete Signals. 
-    Removes nested concatenations and resolves slices along the way. """
-
-    if not len(conc.parts):
-        raise RuntimeError("Concatenation with no parts")
-
-    if all(_flat_concatable(p) for p in conc.parts):
-        return Concat(*[_resolve_sliceable(p) for p in conc.parts])
-
-    if isinstance(conc.parts[0], Concat):
-        # Recursively cover the first element, and all others
-        first = _resolve_concat(conc.parts[0])
-        rest = _resolve_concat(Concat(*conc.parts[1:]))
-        return Concat(*(first.parts + rest.parts))
-
-    if isinstance(conc.parts[0], Slice):
-        # Resolve everything within the Slice to a list of concrete-Signal slices
-        first = _resolve_slice(conc.parts[0])
-        # Pass everything else recursively back to this method
-        rest = _resolve_concat(Concat(*conc.parts[1:]))
-        # And concatenate the two
-        return Concat(*(first + rest.parts))
-
-    # Otherwise peel off as many Signals and concrete-Signal Slices as we can
-    for idx in range(len(conc.parts)):
-        if _flat_concatable(conc.parts[idx]):
-            continue
-        # Hit our first "compound" entry. Split the list here.
-        first = conc.parts[:idx]
-        rest = _resolve_concat(Concat(*conc.parts[idx:]))
-        return Concat(*(first + rest.parts))
-
-    raise RuntimeError("Unable to resolve concatenation")
-
-
-def _flat_concatable(s: Sliceable) -> bool:
-    """ Boolean indication of whether `s` is suitable for flattened Concatenations. 
-    Such objects must be either: 
-    * (a) A Signal, or
-    * (b) A Slice into a Signal
-    Notable exceptions include Concats and nested Slices of Concats and other Slices. """
-    return isinstance(s, Signal) or (
-        isinstance(s, Slice) and isinstance(s.signal, Signal)
-    )
-
-
 class SetList:
     """ A common combination of a hash-set and ordered list of the same items. 
     Used for keeping ordered items while maintaining quick membership testing. """
@@ -1277,81 +738,3 @@ class SetList:
     @property
     def order(self):
         return self.list
-
-
-class ElabPass(Enum):
-    """ Enumerated Elaborator Passes
-    Each has a `value` attribute which is an Elaborator-pass, 
-    and a `name` attribute which is a (Python-enum-style) capitalized name. 
-    Specifying  """
-
-    RUN_GENERATORS = GeneratorElaborator
-    ORPHANAGE = Orphanage
-    IMPLICIT_BUNDLES = ImplicitBundles
-    BUNDLE_CONN_TYPES = BundleConnTypes
-    FLATTEN_BUNDLES = BundleFlattener
-    IMPLICIT_SIGNALS = ImplicitSignals
-    FLATTEN_ARRAYS = ArrayFlattener
-    SIGNAL_CONN_TYPES = SignalConnTypes
-    RESOLVE_SLICES = SliceResolver
-
-    @classmethod
-    def default(cls) -> List["ElabPass"]:
-        """ Return the default ordered Elaborator Passes. """
-        # Returns each in definition order, then a final `Orphanage` test.
-        return list(ElabPass) + [ElabPass.ORPHANAGE]
-
-
-def elab_all(top: Elabables, **kwargs) -> List[Elabable]:
-    """ Elaborate everything we can find - potentially recursively - in `Elabables` `top`. 
-
-    Results are returned in a list, not necessarily reproducing the structure of `top`. 
-    Note the *attributes* of `top` are also generally modified in-place, allowing access to their elaboration results. """
-    # Recursively create a list of all elab-able types in `obj`
-    ls = []
-    _list_elabables_helper(top, ls)
-    # Elaborate each, and return them as a list
-    return [elaborate(top=t, **kwargs) for t in ls]
-
-
-def _list_elabables_helper(obj: Any, accum: list) -> None:
-    """ Recursive helper for hierarchically finding elaborate-able things in `obj`. 
-    Newly-found items are appended to accumulation-list `accum`. """
-    if elabable(obj):
-        accum.append(obj)
-    elif isinstance(obj, list):
-        [_list_elabables_helper(i, accum) for i in obj]
-    elif isinstance(obj, SimpleNamespace):
-        # Note this skips over non-elaboratable items (e.g. names), where the list demands all be suitable.
-        for i in obj.__dict__.values():
-            if isinstance(i, (SimpleNamespace, list)) or elabable(i):
-                _list_elabables_helper(i, accum)
-    else:
-        raise TypeError(f"Attempting Invalid Elaboration of {obj}")
-
-
-def elaborate(
-    top: Elabable,
-    *,
-    ctx: Optional[Context] = None,
-    passes: Optional[List[ElabPass]] = None,
-) -> Module:
-    """ In-Memory Elaboration of Generator or Module `top`. 
-    
-    Optional `passes` lists the ordered `ElabPass`es to run. By default it runs the order specified by `ElabPass.default`. 
-    Note the order of passes is important; many depend upon others to have completed before they can successfully run. 
-
-    Optional `Context` field `ctx` is not yet supported. 
-
-    `elaborate` executes elaboration of a *single* `top` object. 
-    For (plural) combinations of `Elabable` objects, use `elab_all`. 
-    """
-    # Expand default values
-    ctx = ctx or Context()
-    passes = passes or ElabPass.default()
-
-    # Pass `top` through each of our passes, in order
-    res = top
-    for pass_ in passes:
-        res = pass_.value.elaborate(top=res, ctx=ctx)
-    return res
